Collapse the two dashboard layouts into one responsive layout
#20 shipped two separate layouts — a fixed 1024x600 canvas scaled with a CSS
transform for the kiosk, and a reflowed one for everything else — which meant
two sets of components to keep in step for the same 17 panels.

There is now a single Dashboard. The three columns and the card order inside
each are fixed in the component, straight from the design, so no viewport ever
reshuffles them; only the number of columns standing side by side changes:
1 below 640px, 2 up to 1023px, 3 above that at the design's own 238:472:282
proportions.

The 7-inch panel still shows every card without scrolling. Rather than scaling
the whole canvas, one media query (wide and short) drops a set of size tokens —
type scale, padding, gauge diameter, chart and sparkline heights. The
arrangement is untouched there; only the dimensions differ. Checked against a
16-core host with six interfaces and every list full: no overflow, ~28px spare
per column.

Its width trigger is 960px rather than 1024px deliberately. If the cards ever
overflow on a 1024px panel a scrollbar appears and drops the viewport below
1024; with the trigger at 1024 that would switch the type back up and overflow
further, a loop.

Sizes now live in semantic classes (t-label, dash-card, dash-chart, …) in
globals.css instead of being hardcoded per element, which is what lets the two
densities share one set of components.

Removed: KioskDashboard, ResponsiveDashboard, LeftColumn, CenterColumn,
RightColumn, DashboardHeader, AlertBanner, useViewMode, and the ?kiosk=1/0
override that picked between the layouts. Also drops a stale globals.css block
that rescaled the root font size at exactly 1024x600 — a leftover from the old
dashboard that fought the density tokens.
