+++ .env.example
@@ -43,9 +43,5 @@ NEXT_PUBLIC_AUTHOR_NAME=Ruthgyeul
 # --- Kiosk launch script (scripts/run.sh) ----------------------------------
 # Linux user whose Firefox session is killed/relaunched in kiosk mode, and
 # the URL that gets opened.
-#
-# The dashboard picks the fixed 1024x600 layout automatically on a 7-inch
-# panel. Append `?kiosk=1` here only if the browser reports a viewport that
-# makes it fall back to the responsive layout (`?kiosk=0` forces the opposite).
 KIOSK_USER=ruthgyeul
 KIOSK_URL=http://localhost:3000
